Stage 4.8.4b: widen kzopen return type to intptr_t
kzopen returns `(int32_t)kzfs.fil` at five sites — a FILE* cast to
int32_t. On x86_64 the upper 32 bits silently drop; if the FILE*'s
low 32 bits happen to be zero, every `if (!kzopen(...))` caller
would falsely conclude the open failed.

All nine in-tree callers use the return purely as truthiness
(`if (!kzopen(...))`); none treat it as a FILE*. The cleanest fix
is to widen the return type to intptr_t so the truthiness check
sees the full pointer.

Changed:
  - voxlap/voxlap5.h:301  return type int32_t → intptr_t
  - voxlap/kplib.c:2825   function definition matches
  - voxlap/kplib.c:2842/2857/2872/2881/2908  cast (int32_t) → (intptr_t)

Bit-identical on Windows MSVC x86 (intptr_t == int32_t). On Linux
GCC/Clang x86_64 the change clears 5 pointer-to-int-cast warnings
(110/121 remaining). ABI-compatible: callers that read int32_t
return values still see the same low-32-bit pattern they always
have, but no caller does — they all use `!`.

+++ voxlap/kplib.c
@@ -2822,7 +2822,7 @@ void kzsetfil (FILE *fil)
 	kzfs.i = 0;
 }
 
-int32_t kzopen (const char *filnam)
+intptr_t kzopen (const char *filnam)
 {
 	FILE *fil;
 	int32_t i, j, fileoffs, fileleng;
@@ -2839,7 +2839,7 @@ int32_t kzopen (const char *filnam)
 			kzfs.leng = filelength(_fileno(kzfs.fil));
 			kzfs.pos = 0;
 			kzfs.i = 0;
-			return((int32_t)kzfs.fil);
+			return((intptr_t)kzfs.fil);
 		}
 	}
 	if (kzcheckhash(filnam,&zipnam,&fileoffs,&fileleng,&iscomp)) //Then check mounted ZIP/GRP files
@@ -2854,7 +2854,7 @@ int32_t kzopen (const char *filnam)
 			kzfs.leng = fileleng;
 			kzfs.pos = 0;
 			kzfs.i = 0;
-			return((int32_t)kzfs.fil);
+			return((intptr_t)kzfs.fil);
 		}
 		else
 		{
@@ -2869,7 +2869,7 @@ int32_t kzopen (const char *filnam)
 			kzfs.pos = 0;
 			switch(kzfs.comptyp) //Compression method
 			{
-				case 0: kzfs.i = 0; return((int32_t)kzfs.fil);
+				case 0: kzfs.i = 0; return((intptr_t)kzfs.fil);
 				case 8:
 					if (!pnginited) { pnginited = 1; initpngtables(); }
 					kzfs.comptell = 0;
@@ -2878,7 +2878,7 @@ int32_t kzopen (const char *filnam)
 						//WARNING: No file in ZIP can be > 2GB-32K bytes
 					gslidew = 0x7fffffff; //Force reload at beginning
 
-					return((int32_t)kzfs.fil);
+					return((intptr_t)kzfs.fil);
 				default: fclose(kzfs.fil); kzfs.fil = 0; return(0);
 			}
 		}
@@ -2905,7 +2905,7 @@ int32_t kzopen (const char *filnam)
 			kzfs.leng = filelength(_fileno(kzfs.fil));
 			kzfs.pos = 0;
 			kzfs.i = 0;
-			return((int32_t)kzfs.fil);
+			return((intptr_t)kzfs.fil);
 		}
 	}
